+++ backend/config/settings.py
@@ -66,20 +66,23 @@ CORS_ALLOW_HEADERS = [
 INSTALLED_APPS = [
     'django.contrib.admin',
     'django.contrib.auth',
+    'django.contrib.messages',
     'django.contrib.contenttypes',
     'django.contrib.sessions',
-    'django.contrib.messages',
     'django.contrib.staticfiles',
     "corsheaders",
     "rest_framework",
     "rest_framework_simplejwt",
     "users",
-    "posts"
+    "posts",
+    "stories",
+    "messages",
+    "notifications"
 ]
 
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': (
-        'api.authentication.CookiesJWTAuthentication',
+        'users.authentication.CookiesJWTAuthentication',
     ),
     'DEFAULT_PERMISSION_CLASSES': [
         'rest_framework.permissions.IsAuthenticated',
