+++ backend/messages/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class MessagesConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'messages'
+    label = 'user_messages'
