+++ backend/notifications/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class NotifyConfig(AppConfig):
     default_auto_field = 'django.db.models.BigAutoField'
-    name = 'notify'
+    name = 'notifications'
