+++ backend/notifications/serializers.py
@@ -1,0 +1,7 @@
+from rest_framework import serializers
+from notifications.models import Notification
+
+class NotificationSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Notification
+        fields = "__all__"
