+++ backend/notifications/urls.py
@@ -1,0 +1,6 @@
+from django.urls import path
+from notifications.views import *
+
+urlpatterns = [
+    path('notifications/', get_notifications, name='notifications')
+]
