+++ backend/stories/serializers.py
@@ -1,0 +1,8 @@
+from rest_framework import serializers
+from stories.models import Story
+
+
+class StorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Story
+        fields = "__all__"
