+++ backend/users/urls.py
@@ -1,6 +1,6 @@
 from django.urls import path
 from users.token import *
-from users.views import register, logout, get_user, get_users
+from users.views import *
 
 urlpatterns = [
     path('token/', CustomTokenObtainPairView.as_view(), name='token_obtain_pair'),
@@ -8,5 +8,6 @@ urlpatterns = [
     path('register/', register, name='register'),
     path('logout/', logout, name='logout'),
     path('users/', get_users, name='get-users'),
-    path('user/', get_user, name='get-user')
+    path('user/', get_user, name='get-user'),
+    path('follow/<int:user_id>/', follow_user, name='follow-user')
 ]
